Not call Mixpanel's identify if no userId has been set
We shouldn't call Mixpanel's identify method if the userId is not set or not valid. We do that check in the iOS platform here: https://github.com/segment-integrations/analytics-ios-integration-mixpanel/blob/master/Pod/Classes/SEGMixpanelIntegration.m#L44

This was bringing problems when the anonymous Id for a user changed after calling this identify without any user id. We need to be able to set Traits (or Mixpanel superproperties) before we know a user id (that we get later in the process), and later do an alias call to relate both the anonymous id with our user id. However, calling Mixpanel's identify with a null user id before calling alias was causing issues with the anonymous id changing and being the incorrect id.